rename session storage to sessionService
