+++ Maze/Maze/Maze.Server.Commands/CommandFactory/PackageHandlerChain/IPackageHandler.cs
@@ -1,9 +1,6 @@
 ﻿using Maze.Common.MazePackages;
 using Maze.Server.Commands;
 using Maze.Server.Commands.Commands;
-using Maze.Server.Core.Repositories;
-using Maze.Server.Core.ServiceStorage;
-using Maze.Server.Core.SessionStorage;
 using System;
 using System.Linq;
