+++ Maze/Maze/Maze.Server.MazeService.SessionService/Maze.Server.MazeService.SessionService.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{9F6DBD06-7DE8-49BD-9902-094D6E1A103B}</ProjectGuid>
+    <ProjectGuid>{DF334A41-D186-4C97-A536-B1455980A6F6}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Maze.Server.MazeService.SessionStorage</RootNamespace>
-    <AssemblyName>Maze.Server.MazeService.SessionStorage</AssemblyName>
+    <RootNamespace>Maze.Server.MazeService.SessionService</RootNamespace>
+    <AssemblyName>Maze.Server.MazeService.SessionService</AssemblyName>
     <TargetFrameworkVersion>v4.7.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -42,8 +42,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SessionStorage\DumpSessionStorage.cs" />
-    <Compile Include="SessionStorage\ISessionStorage.cs" />
+    <Compile Include="SessionService\DumpSessionService.cs" />
+    <Compile Include="SessionService\ISessionService.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Maze.Common.Model\Maze.Common.Model.csproj">
